fix!: Remove all extra git code
