--- crates/git-branch-stash/src/git/mod.rs
@@ -1,7 +1,0 @@
-mod branches;
-mod protect;
-mod repo;
-
-pub use branches::*;
-pub use protect::*;
-pub use repo::*;
