+++ crates/git-branch-stash/src/lib.rs
@@ -1,10 +1,11 @@
 #![allow(clippy::collapsible_else_if)]
 
 pub mod config;
-pub mod git;
 
+pub use git::*;
 pub use snapshot::*;
 pub use stack::*;
 
+mod git;
 mod snapshot;
 mod stack;
